Cập nhật chức năng CRUD cho bảng Drink và xóa bảng Topping

+++ ManageDrink/src/main/java/com/ManageDrink/repository/ToppingRepository.java
@@ -5,5 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ToppingRepository extends JpaRepository<ToppingEntity,Integer> {
+public interface ToppingRepository extends JpaRepository<ToppingEntity,Long> {
+
+
 }
